backend/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
decorator==5.1.1
defusedxml==0.8.0rc2
Django==3.2
django-extensions==3.2.3
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
exceptiongroup==1.2.0
executing==2.0.1
idna==3.6
iniconfig==2.0.0
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.6
oauthlib==3.2.2
packaging==24.0
parso==0.8.3
pexpect==4.9.0
pluggy==0.13.1
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
stack-data==0.6.3
toml==0.10.2
traitlets==5.14.2
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
